Update README.md
This open-source Android library is aimed at app developers who want to improve the way users are asked for ratings. (as popularised by Circa in http://bit.ly/125IFe6)

It provides full design customisability, is based on a rule engine which is adjustable and/or swappable and supplies a list adapter decorator for automated management of the integrated rating request element in any list.